GLOSSARY-53 Add ability to sync multilang Glossary with deepl glossary
* Fix variable name
* Fix package name
