+++ application-glossary-machine-translation/application-glossary-machine-translation-api/src/main/resources/META-INF/components.txt
@@ -1,2 +1,2 @@
-org.xwiki.contrib.glossary.machineTranslation.TranslationGlossaryScriptService
-org.xwiki.contrib.glossary.machineTranslation.internal.DefaultTranslationGlossaryManager
+org.xwiki.contrib.glossary.machinetranslation.TranslationGlossaryScriptService
+org.xwiki.contrib.glossary.machinetranslation.internal.DefaultTranslationGlossaryManager
